Main/requirements.txt
tensorflow==2.1.0
streamlit==0.55.2
pandas==0.24.2
numpy==1.16.6
mediapipe==0.8.7.3
opencv_contrib_python==4.5.3.56
plotly==4.14.3
matplotlib==2.2.5
keras==2.7.0
Pillow==8.4.0
scikit_learn==1.0.1
